Move patient-sync related classes from '/sync/patient' to '/patient/sync' package
