+++ app/src/androidTest/java/org/resolvetosavelives/red/di/TestAppComponent.kt
@@ -2,7 +2,7 @@ package org.resolvetosavelives.red.di
 
 import dagger.Component
 import org.resolvetosavelives.red.bp.sync.BloodPressureSyncAndroidTest
-import org.resolvetosavelives.red.sync.PatientSyncAndroidTest
+import org.resolvetosavelives.red.patient.PatientSyncAndroidTest
 
 @AppScope
 @Component(modules = [AppModule::class])
