+++ app/src/main/java/org/resolvetosavelives/red/bp/sync/BloodPressureSync.kt
@@ -9,7 +9,7 @@ import io.reactivex.schedulers.Schedulers
 import org.resolvetosavelives.red.bp.BloodPressureRepository
 import org.resolvetosavelives.red.patient.SyncStatus
 import org.resolvetosavelives.red.sync.SyncConfig
-import org.resolvetosavelives.red.sync.patient.DataPushResponse
+import org.resolvetosavelives.red.patient.sync.DataPushResponse
 import org.resolvetosavelives.red.util.Just
 import org.resolvetosavelives.red.util.None
 import org.resolvetosavelives.red.util.Optional
