+++ app/src/main/java/org/resolvetosavelives/red/patient/sync/DataPushResponse.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.sync.patient
+package org.resolvetosavelives.red.patient.sync
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
