+++ app/src/main/java/org/resolvetosavelives/red/patient/sync/PatientSyncApiV1.kt
@@ -1,4 +1,4 @@
-package org.resolvetosavelives.red.sync.patient
+package org.resolvetosavelives.red.patient.sync
 
 import io.reactivex.Single
 import org.threeten.bp.Instant
